requirements.txt
Django==1.9.4
psycopg2==2.6.1
django-extensions==1.5.9
coverage==4.1
uwsgi==2.0.13.1
requests==2.10.0
djangorestframework==3.3.3
supervisor==3.3.0
